requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
httpx==0.28.1
beautifulsoup4==4.13.3
jieba==0.42.1
google-generativeai==0.8.4
scikit-learn==1.6.1
redis==5.2.1
sqlalchemy==2.0.38
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
python-multipart==0.0.20